+++ rpcman-demo-provider/src/main/java/cn/ipman/rpcman/demo/provider/UserServiceImpl.java
@@ -28,7 +28,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public int getId(int id) {
+    public long getId(long id) {
         return id;
     }
